Stop using file id in file naming and send file name in last segment
Previously, the file id was used to control the eventual subdirectory
under which a completed file was stored. In this commit, the stored file
name is now solely determined by the 'file name' field in the message
and not the file id. The file name can also contain subdirectories, in
which case the directory chain would be created (if needed) as part of
the storage operation.

The file id is now used only for associating segment messages to one
another for a stored file and as the name of a working directory in
which file segments are stored while the uploaded file is being
reassembled. The reassembled file is now moved from the working
directory to its final location when the storage operation is
complete.

Previously, the file name of the file to store had to be specified in
the first segment message sent to the DXL fabric for the file. This
commit changes the behavior to send the file name in the last segment of
the file. This lays some groundwork for a future ability to be able to
handle the case where segments are sent somewhat out of order, in which
case the file name might otherwise not have been available when a
segment after the first one for a file is processed first.