Attempt to calculate total_segments for stream request if not specified
Previously, if call made to the `send_file_from_stream_request` included
the `stream_size` and `max_segment_size` parameters but not the
`total_segments` parameter, the `total_segments` parameter was passed
through to the callback with a value of `None`. With the changes in this
commit, the value for the `total_segments` parameter is calculated
automatically from the `stream_size` and `max_segments_size` parameters.

The ability to specify the `total_segments` parameter has been left in
place, mainly for backward compatibility. As with prior commits, if the
number of segments sent reaches the `total_segments` value before the
number of bytes read reaches the `stream_size` value, no further bytes
will be sent.